+++ airflow/providers/amazon/aws/hooks/sagemaker.py
@@ -35,7 +35,7 @@ from airflow.exceptions import AirflowException
 from airflow.providers.amazon.aws.hooks.base_aws import AwsBaseHook
 from airflow.providers.amazon.aws.hooks.logs import AwsLogsHook
 from airflow.providers.amazon.aws.hooks.s3 import S3Hook
-from airflow.providers.amazon.aws.utils.aws_api import format_tags
+from airflow.providers.amazon.aws.utils.tags import format_tags
 from airflow.utils import timezone
