+++ airflow/providers/amazon/aws/utils/tags.py
